moved the deletion of nodes to be in the scene object. When deleting nodes, the remaining nodes indexes get reordered
